# docs/api/integrated-api-openapi.json
{
    "components": {
        "schemas": {
            "ErrorResponse": {
                "description": "\uacf5\ud1b5 \uc5d0\ub7ec \uc751\ub2f5",
                "properties": {
                    "code": {
                        "description": "\uc5d0\ub7ec \ucf54\ub4dc",
                        "example": "BE40101",
                        "type": "string"
                    },
                    "message": {
                        "description": "\uc5d0\ub7ec \uba54\uc2dc\uc9c0",
                        "example": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                        "type": "string"
                    },
                    "status": {
                        "description": "HTTP \uc0c1\ud0dc \ucf54\ub4dc",
                        "example": 401,
                        "format": "int32",
                        "type": "integer"
                    },
                    "timestamp": {
                        "description": "\uc5d0\ub7ec \ubc1c\uc0dd \uc2dc\uac01",
                        "format": "date-time",
                        "type": "string"
                    }
                }
            },
...
            "com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Bootstrap": {
                "description": "Elasticsearch alias \ubd80\ud2b8\uc2a4\ud2b8\ub7a9 \uacb0\uacfc",
                "properties": {
                    "alias": {
                        "description": "alias \uc774\ub984",
                        "type": "string"
                    },
                    "created": {
                        "description": "\uc774\ubc88 \ud638\ucd9c\uc5d0\uc11c \uc778\ub371\uc2a4\ub97c \uc0c8\ub85c \uc0dd\uc131\ud588\ub294\uc9c0 \uc5ec\ubd80",
                        "type": "boolean"
                    },
                    "index": {
                        "description": "\uc0dd\uc131\ub41c \ub610\ub294 \uae30\uc874 \uc778\ub371\uc2a4 \uc774\ub984",
                        "type": "string"
                    }
                }
            },
            "com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Reindex": {
                "description": "Elasticsearch \uc7ac\uc0c9\uc778 \uacb0\uacfc",
                "properties": {
                    "alias": {
                        "description": "alias \uc774\ub984",
                        "type": "string"
                    },
                    "documents": {
                        "description": "\uc7ac\uc0c9\uc778\ub41c \ubb38\uc11c \uc218",
                        "format": "int64",
                        "type": "integer"
                    },
                    "new_index": {
                        "description": "\uc7ac\uc0c9\uc778 \ud6c4 alias\uac00 \uac00\ub9ac\ud0a4\ub294 \uc0c8 \uc778\ub371\uc2a4 \uc774\ub984",
                        "type": "string"
                    },
                    "source_index": {
                        "description": "\uc7ac\uc0c9\uc778 \uc6d0\ubcf8 \uc778\ub371\uc2a4 \uc774\ub984",
                        "type": "string"
                    }
                }
            },
            "com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Status": {
                "description": "Elasticsearch alias \uc0c1\ud0dc \uc870\ud68c \uacb0\uacfc",
                "properties": {
                    "alias": {
                        "description": "alias \uc774\ub984",
                        "example": "techblog-posts",
                        "type": "string"
                    },
                    "current_index": {
                        "description": "alias\uac00 \uac00\ub9ac\ud0a4\ub294 \ud604\uc7ac \uc778\ub371\uc2a4 \uc774\ub984",
                        "example": "techblog-posts-v1",
                        "type": "string"
                    },
                    "exists": {
                        "description": "alias \uc874\uc7ac \uc5ec\ubd80",
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "bearerFormat": "JWT",
                "description": "user-service \ub85c\uadf8\uc778 \ub610\ub294 \uad6c\uae00 OAuth \ucf5c\ubc31\uc5d0\uc11c \ubc1c\uae09\ubc1b\uc740 access token",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "integrated-service REST API",
        "title": "Integrated-Service API",
        "version": "v1"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/index/v1/elasticsearch/_bootstrap": {
            "post": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "bootstrap",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Bootstrap"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc778\ub371\uc2a4 \ubd80\ud2b8\uc2a4\ud2b8\ub7a9 (alias \uc5c6\uc73c\uba74 \ubb3c\ub9ac \uc778\ub371\uc2a4 + alias \uc0dd\uc131)",
                "tags": [
                    "07. Elasticsearch \uc778\ub371\uc2a4 \uad00\ub9ac API"
                ]
            }
        },
        "/index/v1/elasticsearch/_reindex": {
            "post": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "reindex",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Reindex"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "\uc7ac\uc0c9\uc778 \uc2e4\ud589 (\uc0c8 \ubb3c\ub9ac \uc778\ub371\uc2a4 \uc0dd\uc131 \u2192 reindex \u2192 alias \uc6d0\uc790\uc801 \uc2a4\uc651)",
                "tags": [
                    "07. Elasticsearch \uc778\ub371\uc2a4 \uad00\ub9ac API"
                ]
            }
        },
        "/index/v1/elasticsearch/status": {
            "get": {
                "description": "\uc6b4\uc601\uc790 \uc804\uc6a9 API",
                "operationId": "getStatus",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/com.backend.integratedapi.elasticsearchindex.controller.dto.ElasticsearchIndexReadDto.Status"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE40001",
                                    "message": "\uc785\ub825 \ub370\uc774\ud130\uc5d0 \ubb38\uc81c\uac00 \uc788\uc2b5\ub2c8\ub2e4.",
                                    "status": 400,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uc694\uccad \uac12\uc774 \uc62c\ubc14\ub974\uc9c0 \uc54a\uc74c"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40101",
                                    "message": "\uc778\uc99d\ub418\uc9c0 \uc54a\uc740 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 401,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ud1a0\ud070\uc774 \uc5c6\uac70\ub098 \uc720\ud6a8\ud558\uc9c0 \uc54a\uc74c"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "BE40301",
                                    "message": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790\uc785\ub2c8\ub2e4.",
                                    "status": 403,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\uad8c\ud55c\uc774 \uc5c6\ub294 \uc0ac\uc6a9\uc790"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": "FE50001",
                                    "message": "\uc11c\ubc84 \ub0b4\ubd80 \uc624\ub958\uac00 \ubc1c\uc0dd\ud588\uc2b5\ub2c8\ub2e4.",
                                    "status": 500,
                                    "timestamp": "2026-08-09T12:00:00+09:00"
                                },
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        },
                        "description": "\ucc98\ub9ac\ub418\uc9c0 \uc54a\uc740 \uc11c\ubc84 \uc624\ub958"
                    }
                },
                "summary": "alias \ud604\uc7ac \uc0c1\ud0dc \uc870\ud68c (alias \u2192 \ubb3c\ub9ac \uc778\ub371\uc2a4)",
                "tags": [
                    "07. Elasticsearch \uc778\ub371\uc2a4 \uad00\ub9ac API"
                ]
            }
        },
...
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "servers": [
        {
            "description": "Generated server url",
            "url": "http://localhost:8081"
        }
    ],
...
}